Add files via upload
Added the Derivative Calculator. Right now it is a work in progress, as it can only take the derivative of single variable functions with no exponent that are being multiplied by a single digit. In the final version, the user will be able to enter a function in the first line of code and get the derivative of the function.